Move datastore Drush.php to DatastoreCommands
