--- modules/dkan_datastore/drush.services.yml
@@ -1,14 +1,0 @@
-services:
-  dkan_datastore:commands:
-    class: \Drupal\dkan_datastore\Drush
-    arguments:
-      - '@dkan.metastore.service'
-      - '@dkan.datastore.service'
-      - '@dkan.datastore.service.resource_localizer'
-      - '@dkan.metastore.resource_mapper'
-      - '@dkan.datastore.import_info_list'
-      - '@dkan.datastore.post_import_result_factory'
-      - '@dkan.datastore.lookup'
-      - '@state'
-    tags:
-      - { name: drush.command }
