+++ modules/dkan_metastore/dkan_metastore.services.yml
@@ -7,6 +7,7 @@ services:
       - '@dkan.metastore.valid_metadata'
       - '@dkan.common.logger_channel'
       - '@event_dispatcher'
+  Drupal\dkan_metastore\MetastoreService: '@dkan.metastore.service'
 
   dkan.metastore.lifecycle:
     class: \Drupal\dkan_metastore\LifeCycle\LifeCycle
@@ -77,6 +78,7 @@ services:
     arguments:
       - '@entity_type.manager'
       - '@event_dispatcher'
+  Drupal\dkan_metastore\ResourceMapper: '@dkan.metastore.resource_mapper'
 
   dkan.metastore.event_subscriber:
     class: \Drupal\dkan_metastore\EventSubscriber\MetastoreSubscriber
